chore(project): add pyproject.toml
This commit adds a pyproject.toml for kepler-model-server based on
guidelines for creating python packaging[1]. The build backend used is
hatch since hatch helps manage a lot of packaging and enviroment related
chores easily.

`hatch shell` or (`pip install .`) installs a `model-server`, and an
estimator script. Minimal changes are made to the files to get the
script running.

[1]: https://packaging.python.org/en/latest/guides/writing-pyproject-toml/

Signed-off-by: Sunil Thaha <sthaha@redhat.com>